+++ fe2o3-amqp-types/src/performatives/end.rs
@@ -5,7 +5,7 @@ use crate::definitions::Error;
 /// <type name="end" class="composite" source="list" provides="frame">
 /// <descriptor name="amqp:end:list" code="0x00000000:0x00000017"/>
 /// </type>
-#[derive(Debug, Clone, DeserializeComposite, SerializeComposite)]
+#[derive(Debug, Clone, PartialEq, Eq, DeserializeComposite, SerializeComposite)]
 #[amqp_contract(
     name = "amqp:end:list",
     code = "0x0000_0000:0x0000_0017",